app/requirements.txt
python-decouple==3.8
requests==2.32.3
spitch==1.16.0
